Validate custom components service.yaml
